requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4

djangorestframework~=3.14.0